requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0
tqdm>=4.64.0
scipy>=1.7.0
tensorboard>=2.10.0
pytest>=7.0.0
